Code update - 2022-09-30

+++ services/scaling-service/src/services/vultr/vultr.service.ts
@@ -83,16 +83,17 @@ export class VultrService {
 
   async terminateInstance(instanceId): Promise<void> {
     try {
-      this.logger.debug(`terminate instance with id: ${instanceId}`);
+      if (instanceId) {
+        this.logger.debug(`terminate instance with id: ${instanceId}`);
 
-      await sendHttpRequest({
-        url: getInstanceUrl({ instanceId }),
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
-        },
-      });
+        await sendHttpRequest({
+          url: getInstanceUrl({ instanceId }),
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+          },
+        });
+      }
     } catch (e) {
       this.logger.debug(`error terminating instance with id: ${instanceId}`);
       this.logger.error(e.message);
